1. Initial version of uniprotkb_owlnets.py, which adds data from the UNIPROTKB protein database. 2. Enhancements/fixes to OWLNETS-UMLS-GRAPH-12.py:
a. Special handling of HGNC concepts as object nodes in relationships, to account for the fact that HGNC codes are stored in the knowledge graph with 2 delimiters.

b. Fix to find all inverse relationships in RO, not just those with an obvious "inverseOf" names.

c. Enhancement to address case (such as UNIPROTKB-HGNC) in which relationships are between the new ontology and an external ontology (one not defined in node_metadata.txt)